refactor: isolate release and platform capabilities
